minor: split tokenizer_info.cc
