+++ cpp/tokenizer_info.cc
@@ -3,92 +3,23 @@
  * \file xgrammar/tokenizer_info.cc
  */
 
+#include "xgrammar/tokenizer_info.h"
+
 #include <picojson.h>
-#include <xgrammar/tokenizer_info.h>
 
 #include <array>
 #include <memory>
 #include <optional>
 #include <string>
 #include <unordered_map>
-#include <unordered_set>
 #include <vector>
 
 #include "support/encoding.h"
+#include "support/json_serialize.h"
 #include "support/logging.h"
+#include "tokenizer_internal.h"
 
 namespace xgrammar {
-
-class TokenizerInfo::Impl {
- public:
-  Impl(
-      const std::vector<std::string>& encoded_vocab,
-      VocabType vocab_type,
-      std::optional<int> vocab_size,
-      std::optional<std::vector<int32_t>> stop_token_ids,
-      bool add_prefix_space
-  );
-
-  VocabType GetVocabType() const { return vocab_type_; }
-  bool GetAddPrefixSpace() const { return add_prefix_space_; }
-  int GetVocabSize() const { return vocab_size_; }
-  const std::vector<std::string>& GetDecodedVocab() { return decoded_vocab_; }
-  const std::vector<int32_t>& GetStopTokenIds() const { return stop_token_ids_; }
-  const std::vector<int32_t>& GetSpecialTokenIds() const { return special_token_ids_; }
-  const std::vector<std::pair<int32_t, std::string>>& GetSortedDecodedVocab() const {
-    return sorted_decoded_vocab_;
-  }
-
-  std::string DumpMetadata() const;
-
-  static std::shared_ptr<Impl> FromVocabAndMetadata(
-      const std::vector<std::string>& encoded_vocab, const std::string& metadata
-  );
-
-  static std::string DetectMetadataFromHF(const std::string& backend_str);
-
- private:
-  static bool IsSpecialToken(const std::string& decoded_token);
-
-  /*! \brief The vocabulary type. */
-  VocabType vocab_type_;
-  /*! \brief The size of the vocabulary. */
-  int vocab_size_;
-  /*! \brief Whether to add prefix space. */
-  bool add_prefix_space_;
-
-  /*! \brief The vocabulary. Special tokens are included. */
-  std::vector<std::string> decoded_vocab_;
-  /*! \brief All (id, token) pairs sorted in lexicographic order. This sorting is done to
-   * maximize prefix reuse during matching. Special tokens and stop tokens are not included. */
-  std::vector<std::pair<int32_t, std::string>> sorted_decoded_vocab_;
-  /*! \brief The stop tokens. When the GrammarMatcher can reach the end of the grammar,
-   * stop tokens can be accepted. */
-  std::vector<int32_t> stop_token_ids_;
-  /*! \brief The special tokens. These tokens are ignored (masked out) during the grammar-guided
-   * generation. */
-  std::vector<int32_t> special_token_ids_;
-
-  /*!
-   * \brief The tokens used to detect stop tokens from the vocabulary.
-   *
-   * LLaMA2: </s>
-   * LLaMA3: <|end_of_text|>, <|eot_id|>
-   * Phi-2: <|endoftext|>
-   * Gemma: <eos>, <end_of_turn>
-   * DeepSeek-V2: <｜end▁of▁sentence｜>
-   */
-  inline static const std::unordered_set<std::string> DETECTION_STOP_TOKENS = {
-      "</s>",
-      "<|end_of_text|>",
-      "<|eot_id|>",
-      "<|endoftext|>",
-      "<eos>",
-      "<|eos|>",
-      "<end_of_turn>",
-      "<｜end▁of▁sentence｜>"
-  };
-};
 
 /************* Token decoders: ByteFallback and ByteLevel *************/
 
@@ -366,17 +297,7 @@ TokenizerInfo::Impl::Impl(
 }
 
 std::string TokenizerInfo::Impl::DumpMetadata() const {
-  picojson::object obj;
-  obj["vocab_type"] = picojson::value(static_cast<int64_t>(vocab_type_));
-  obj["vocab_size"] = picojson::value(static_cast<int64_t>(vocab_size_));
-  obj["add_prefix_space"] = picojson::value(add_prefix_space_);
-  picojson::array stop_token_ids_array;
-  for (auto id : stop_token_ids_) {
-    stop_token_ids_array.push_back(picojson::value(static_cast<int64_t>(id)));
-  }
-  obj["stop_token_ids"] = picojson::value(stop_token_ids_array);
-
-  return picojson::value(obj).serialize(false);
+  return AutoJSONSerialize(*this).serialize(false);
 }
 
 std::shared_ptr<TokenizerInfo::Impl> TokenizerInfo::Impl::FromVocabAndMetadata(
